Add instruction emulation via POWERPC_EXCP_HV_EMU
On AIX, binaries built as 32-bit can still make use of newer ISA
extensions that are not supported by older 32-bit processors.
For example, the `cmpb` instruction was introduced in PowerISA v2.05
(POWER6+), and may appear in such binaries.

This leaves me with two possible approaches:

a) Use a 64-bit processor. This would normally be the logical choice,
   since 64-bit PowerPC CPUs are capable of running 32-bit binaries.
   Unfortunately, PPC64 support in Unicorn is still quite buggy as of
   13-Dec-2025 (v2.0.4). Even when configuring a POWER7 CPU, I am unable
   to execute ISA v2.05 instructions reliably.

b) Use a 32-bit processor together with POWERPC_EXCP_HV_EMU. When a
   PowerPC CPU encounters an unsupported instruction, this exception
   is raised, allowing the operating system to “polyfill” (emulate)
   the missing instruction in software.

Given the current unreliability of PPC64, we are unfortunately forced
to take the second approach. Hopefully, the number of instructions
that require emulation will be small.